[iOS] Update RequestDesktop flag
As the feature won't be launching, this CL updates the code using the
current flag to request MOBILE by default but leaving the architecture
in place if we want to reuse it in the future.

This CL makes the flag refactoring only.
It also removes mobile-google-srp as it is no longer needed.

Bug: 1251138
Change-Id: I0f5f1f4433cfdad4d884e21974dbcd3d57fb75eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3171060
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923782}
NOKEYCHECK=True
GitOrigin-RevId: 2fc86ca5689f92486c5c5068e51def023f892c6e