Remove NOTIFICATION_NAV_ENTRY_PENDING.
Refactor ChromeOmniboxNavigationObserver (which was the only user of
NOTIFICATION_NAV_ENTRY_PENDING) and remove said notification.

Instead of listening to the NotificationService notification to start
the fetch, pass NavigationRequest explicitly (which is possible as it's
the OmniboxEditModel who opens it).

Now ChromeOmniboxNavigationObserver is refcounted and is kept for as
long as either:
- NavigationRequest hasn't finished;
- Alternative URL fetcher hasn't finished.

When both are done and if the fetcher has succeeded,
ChromeOmniboxNavigationObserver's destructor will show the alternative
navigation infobar status.

The following changes to the content/public were made to allow that:
- Allow specifying body in NavigationSimulator (as 404 handling depends
based on whether the body is null or not).

The following changes were originally a part of this patch, but were
landed separately:
- Return WeakPtr<NavigationHandle> to created navigation from
NavigationController::LoadURL (crrev.com/c/2829263).
- Add NavigationHandleUserData to allow associating data with ongoing
navigation (was landed as a part of crrev.com/c/2721190).

R=jdonnelly@chromium.org,avi@chromium.org,jdonnelly@chromium.org
CC=​​​​alexmos@chromium.org,creis@chromium.org,rakina@chromium.org
BUG=1174759

Change-Id: I43a528d04df06a6fd02fa4cde41a9cf6e21f98cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2702335
Commit-Queue: Alexander Timin <altimin@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924303}
NOKEYCHECK=True
GitOrigin-RevId: 66dab131ee7a97abe834d1c6f7ef2e03a0a5c961