[iOS] Fix handling of repeated queries in SafeBrowsingQueryManager
When there are repeated Safe Browsing queries for the same URL
(which can happen when a page has multiple iframes with the same
URL), SafeBrowsingQueryManager::StoreUnsafeResource assumes
that responses will arrive in the same order as the queries are made.

This assumption doesn't always hold in practice when Safe Browsing has
to perform network requests to respond to a query.

This CL removes this assumption. Instead of assuming that a given
UnsafeResource corresponds to the first query with a matching URL,
StoreUnsafeResource now associates the UnsafeResource with all
pending queries that have a matching URL.

Change-Id: Id82db2c41fc1d2080e7c0bfb931f0c8ec282abbb
Bug: 1205980
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3176801
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924748}
NOKEYCHECK=True
GitOrigin-RevId: b1a5b57e3d93a9160e5a0ea9ab30323552695692