[ios] Pass WebStatePolicyDecider::RequestInfo by value
The structure is small enough (it fits in a 64-bit integer) that
the compiler can generate more optimal code when it is passed by
value (as it does not have to spill the value to memory).

See https://godbolt.org/z/WvWbz8563 for comparison of generated
code for arm64 (similar results for x86_64).

Bug: none
Change-Id: I944665620a7f68074053cc2339f7b08cee5c7271
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3182122
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#925241}
NOKEYCHECK=True
GitOrigin-RevId: f9834fcce1b990a9b26455e1573d77736b88a16f