[iOS] Add initial Overflow Menu view
For the view, it is in SwiftUI, currently just a temporary red square.
The provider is necessary because SwiftUI types are not exposed to
Objective C, so we need a Swift wrapper to expose a plain
UIViewController. Similarly, free Swift functions are also not exposed
to Objective C, so it needs to be wrapped in a class.

Bug: 1253367
Change-Id: Id0ccff3b7b35394e2711ab30963cb6c90380cfda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3186852
Reviewed-by: David Jean <djean@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#925755}
NOKEYCHECK=True
GitOrigin-RevId: 659eb6a5b0f3c444cd8005525b6a3be4549797a4