[iOS] Defer granting sync consent when opening advanced settings
This CL defers granting the sync consent in advanced sync settings. The user is only signed-in when the settings are opened.
Encryption is made available in the advanced Settings flow for users that do or do not have a password.
On top of that, before this CL the user was signed out when going from signed-in to syncing. This is no longer the case.

Bug: 1242876, 1253384
Change-Id: I09492f4eb5480001c135693730b4bf046e2b6eff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3151830
Commit-Queue: Maria Lavinia Nedelea <lavinian@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Jérôme <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926704}
NOKEYCHECK=True
GitOrigin-RevId: 9a7ecc1b425818ba907e1cabaa3163eab84c6933