[ios] Stop checking for iOS dictation unicode when updating omnibox.
The method, ShouldIgnoreUserInputDueToPendingVoiceSearch, was added in
OmniboxViewIOS to fix a crash in iOS 7, which appears to be fixed now.
The underlying issue was that modifying the omnibox while dictation was
processing would cause a crash. Since this code no longer seems to be
necessary, we are going to try always returning false from this method.
To be safe, we're putting it behind a Finch flag which defaults to
enabled.

Bug: 1253380

Change-Id: Ia22632373c628be55e891722cae31f3734b26b91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3192611
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#926765}
NOKEYCHECK=True
GitOrigin-RevId: e2aa0f6108afed998c392e96a1fed873b5a1c86b