[iOS] Defer granting sync consent when opening advanced settings
This CL defers granting the sync consent in advanced sync settings. The user is only signed-in when the settings are opened.
Encryption is made available in the advanced Settings flow for users that do or do not have a password.
On top of that, before this CL the user was signed out when going from signed-in to syncing. This is no longer the case.

This patch is a reland of crrev.com/c/3151830
This patch adds a fix in AdvancedSettingsSigninTestCase/testSigninWithPassword to scroll the advanced settings view for small screens (like iPhone SE first generation).

Bug: 1242876, 1253384
Change-Id: Id790f489060b3b88b29a0c9d83d5a7e8982b3f92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198250
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Jérôme <jlebel@chromium.org>
Commit-Queue: Maria Lavinia Nedelea <lavinian@google.com>
Cr-Commit-Position: refs/heads/main@{#927139}
NOKEYCHECK=True
GitOrigin-RevId: fd690c3515c834ec3ece59fad826572bf624288f