[iOS] Add a flag that makes the UA string use version 100
This adds a feature flag and corresponding entry in chrome://flags
that changes the major version in the User-Agent string to 100.

This is intended to allow testing whether sites break with 3-digit
version numbers.

An identical flag has already been added for non-iOS in
crrev.com/c/3167270, and this CL ports the logic added there to iOS.

Bug: 1252810
Change-Id: Ic24023360de7e5fb8e8bc3af053052105dd0ace1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3197470
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ali Beyad <abeyad@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927195}
NOKEYCHECK=True
GitOrigin-RevId: 02a9b7b432cce69b6f1b60ec6d225241c153c2ed