[iOS] CPE: Move focus to save button after generating a password
This helps alert Voiceover users that something has changed when a
password is generated.

There are also 2 other small acessibility fixes in this CL based on
recommendations from the accessibility review. First, the placeholder
had too low contrast with the background, so was unreadable for some
users. This increases the contrast. Second, there was a small bug with
setting the accessibility label on the hide/show password icon,
resulting in the label not being set on initial load of the view.

Bug: 1254797
Change-Id: I3d7e2fb765bb4d8665290f672225ec0e0f0cbdf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198251
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927240}
NOKEYCHECK=True
GitOrigin-RevId: 74e08fde05b3e3975995788788872fee5b850759