[ios] Updates the testrunner to force an arm64 simulator if needed.
Python is currently invoked as an x86_64 binary on arm64 Macs, which
ends up invoking the simulator as an x86_64 binary as well. This causes
test failures because our tests only have an arm64 slice. We can use the
"arch -arch arm64" command to force xcodebuild to run as an arm64
binary, and we use "sysctl sysctl.proc_translated" to detect when it is
necessary to force arm64.

BUG=1250511

Change-Id: If0fc590f501434a77192f9fbefa8476af23f8aac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3192655
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927663}
NOKEYCHECK=True
GitOrigin-RevId: 6478b90d96746371c1e84aba394356d9764b91d4