Create public API getPriceDrop() for a WebState's URL
A public API getPriceDrop() is created which determines if a
WebState's URL contains a price drop or not and if it does, what
the price drop is. This is achieved by creating a TabHelper and
integrating with OptimizationGuide which is the bridge between
Chrome and price data acquisition. This API will power our new
price drop feature where price drops are displayed in the Tab
Switching UI.

Bug: 1245021
Change-Id: I56f7b6432b7437834aacd26c1994021a96388160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198533
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927924}
NOKEYCHECK=True
GitOrigin-RevId: b8fa5447bd92f2cf7c5f9f4b3b24630e8146a9be