[iOS] Add model classes for overflow menu
This is a first-pass at model classes for the overflow menu. Currently,
only a few properties of the model objects are variable, but that may
change. It's also missing some properties (ones to show unread, or
problem, or new markers), but this should be enough to get started
writing the actual views.

Bug: 1254376
Change-Id: Id9971de24db9daa9f79fd476f3dd939c3f6365ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3200070
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928186}
NOKEYCHECK=True
GitOrigin-RevId: 4152426603398018f939f941d2f3be69ca0f377a