[iOS][Passwords] Validate for duplicate existing credentials on every keystroke
With this CL, when the user fills in the add credential form, on every
keystroke, a background check is done to ensure that there is no
existing credential with the same username/website combination.
It introduces AddPasswordDetailsConsumer to inform the view controller
by the mediator of the de-dupe status.
A subsequent CL would optimize the logic to ensure that the duplicate
credential validation is run only when the Website text is valid.

Bug: 1226006
Change-Id: Iee3e43bceff063e9010a4cb3bb0df2e10f71e6f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3192681
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Cr-Commit-Position: refs/heads/main@{#928598}
NOKEYCHECK=True
GitOrigin-RevId: 6d09c94d3fa85ec6825c217fec0cbc68ab6b363d