Set page load tokens in real time URL check and password protection.
Page load tokens should meet the following criteria:
  1) Tokens are the same for requests on the same page.
  2) Tokens are different for requests on different pages.

In this CL, tokens are mapped to the hostname of the current mainframe
URL and are refreshed during mainframe real time URL check. This will
cause some false matching if real time URL check is disabled, and the
user has typed in their passwords multiple times on URLs with the same
hostname. This is acceptable for now, because password protection
request is the only request that is sending tokens, which cannot be
linked. In the future, we will implement a more comprehensive approach
to map to pages more explicitly.

See http://go/chrome-sb-pageload-tokens for more discussion.

The feature is controlled by a flag kSafeBrowsingPageLoadToken. The flag
is disabled by default.

Bug: 1240403
Change-Id: I60283b74017c2a1250ed951389a690146b64aa37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3200182
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928731}
NOKEYCHECK=True
GitOrigin-RevId: 83872c8bf9bed4b67d6551a5ad079cb1f1244d92