[iOS][CPE Promo] Add 'Passwords In Other Apps' item in Password Settings
This CL adds a new item "Passwords in Other Apps" to Chrome Settings -->
Passwords, if the feature
"chrome://flags#credential-provider-extension-promo" is turned on.

Ideally, this item should show whether the Chrome passwords autofill
feature is turned on or not, and it should click to a new view that
serves as the promotional view for the Chrome passwords autofill feature.

Bug: 1252105
Change-Id: I2c68eade652ecf26465dc6cc72c3b65687777130
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3206266
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928782}
NOKEYCHECK=True
GitOrigin-RevId: a69dac161020f4f270378bbfb88ea011a67733b1