Fix an edge case in the Extended Variations Safe Mode experiment.
Temporarily disables the experiment on iOS since the fix requires some
EG tests to be updated. This will be done in a follow-up CL.

When the CleanExitBeacon is created, the beacon file contents are read
and used regardless of whether the client is in the group that uses this
file. It is possible for a client to have the new file and to not be in
the SignalAndWriteViaFileUtil group if the client was previously in that
group and then later switched groups, e.g. via kResetVariationState.

This change fixes the edge case by moving experiment group assignment
from the VariationsFieldTrialCreator to the CleanExitBeacon ctor. This
requires plumbing the channel through to the CleanExitBeacon because
this Variations Safe Mode client-side field trial is active on only
certain channels.

The initial attempt at this change was crrev.com/c/3120545.

Bug: 1244334
Change-Id: I6e714eb9c518369c4546b5b5676aec6bd4fb0c40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3171574
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#929595}
NOKEYCHECK=True
GitOrigin-RevId: 1b4ccc8e01f324502d417e7f46f8aa1c62ec48c4