[iOS][Passwords] Remove Code for Alert from Add Credential flow
This CL removes the code for showing the alert in the add credential
flow when an already existing profile exists with the same
username/website data. The alert is not longer required instead a
section alert would be shown to the user.

Bug: 1226006
Change-Id: I08c32350b5d9423009163c2c865314644c3bb5bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199783
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#929842}
NOKEYCHECK=True
GitOrigin-RevId: 8fed763d4ff5a66c3135758fb455392f640f49b4