[ios] Remove code duplication in WKNavigationDelegate impl
The method -webView:decidePolicyForNavigationAction:decisionHandler:
is not longer needed now that //ios/web requires iOS 13.0 or later.

-webView:decidePolicyForNavigationAction:preferences:decisionHandler:
is the only method that needs to be implemented. As the two methods
only diverged in the decisionHandler interface, merge the old method
in the new one and introduce a local variable to store a wrapped
version of the decisionHandler.

Bug: none
Change-Id: I65caf1809a7e5b78e2eb645ca5719f173a34b225
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3194335
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#930163}
NOKEYCHECK=True
GitOrigin-RevId: 51fc4f2d67b4d50a90c3512c87e3e74f2862a184