Price Drop Chip for adding to Tab Switching UI
This CL adds a price drop chip to be annotated on top of the
screenshot in the Tab Switching UI. It is controlled by a centralized
gate which is currently set to false (so it is not shown). There will
be follow up CLs to enable it for non-incognito, MSBB users with the
flag turned on.

Bug: 1245027
Change-Id: I8fb66105053ddba3436a01ebad9229c8541f4794
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3212310
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#930206}
NOKEYCHECK=True
GitOrigin-RevId: 842fafaf5bd60d4c1a0523b2cd6a7a98354c9344