[ios] Make //services/network dependency on WebRTC optional
Chrome on iOS does not use WebRTC, however //services/network has
a dependency on WebRTC and through it on coded libraries (such as
libvpx, libaom, ...).

Those dependency are unused by Chrome on iOS but they still need
to be built and linked into the final binary which increase time
of compilation and binary size.

Additionally, those libraries causes problem when trying to build
Chrome on iOS (or some subset) for the mac Catalyst environment
(a way to compiled iOS app so that they can run natively on macOS)
or as fat binaries (as the library select optimised implementation
based on the cpu).

Since the code is not needed by Chrome on iOS, add a gn variable
to control whether it is enabled or not. The variable defaults to
`true` as some remoting needs WebRTC (and is enabled by default).

Configure the Chrome on iOS bots to set the variable to `false`.
Disable the feature for Chrome on iOS developers via setup-gn.py.

Move //jingle:fake_ssl_socket to //jingle/glue:fake_ssl_socket to
avoid building //third_party/webrtc_overrides:webrtc_component if
ninja is invoked without parameter (gn generate a ninja target for
all gn target that are defined even if they are not reachable from
the "root" target). By moving this target, jingle/BUILD.gn is not
loaded, and none of its target nor their dependency loaded.

This reduce the number of step to build "all" by ~6000 (or a bit
over 8% less steps).

Bug: 1257509, 1257296
Cq-Include-Trybots: luci.chromium.try:ios-device
Change-Id: I22b11cf94c31a4ab8403d30e47ddeee77e2cd83a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3211504
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#930460}
NOKEYCHECK=True
GitOrigin-RevId: 1561e0d97d94997cefb6c40312803e4f395852d9