Record a set of pseudo metrics for data validation
Add a set of UMA.Pseudo{Metrics} which are recorded alongside with their
mirrored histograms. The sample values of these metrics equal to the
values of the real metrics multiplied by a factor that is controlled by
a field trial experiment.

This is used for data validation. It's not meant to measure any state of
our clients and shouldn't have impact on them.

Bug: 1257484
Change-Id: I59206324249cadc3e547cd1dcffa74f9d5ec7710
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3211366
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Weilun Shi <sweilun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#931243}
NOKEYCHECK=True
GitOrigin-RevId: ef3dd1af677e65e426b2df90c2715ce338e92353