[iOS] AuthenticationService needs to skip sync consent if already set
When sync is disabled by enterprise policy, the sync consent is not
removed.
Once the enterprise policy is removed, to turn sync back on,
AuthenticationService needs to skip the sync consent if it is already
set in order turn on sync.

Fixed: 1256791
Change-Id: I7152272b1c0204e8857f29245e7d6b8849973878
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3220478
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Commit-Queue: Jérôme <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#931433}
NOKEYCHECK=True
GitOrigin-RevId: a4e6a243909ca21bf67a1600918e789fcaee7d8f