Extract tab matcher logic to dedicated translation units.
Near-verbatim move of logic to a dedicated source files. This is meant
to help to declutter the ChromeAutocompleteProviderClient, allowing
us to add a bit of extra logic to the Android implementation to
prefetch and cache the open tabs to significantly accelerate the
SwitchToTab performance.

There are no changes to existing logic, the code is moved with minimum
modifications allowing clean compilation.

Some of the code might look questionable (specifically the ifdefs in
the desktop impl file) -- please note that this is because of existing
code relying on a mechanism that may not be needed at all; deleting
this code would cause a larger diff in this CL, making the change hard
to review. The objective of this change is to extract the logic with
no functional changes.

Bug: 1176768
Change-Id: Ia63453f1da44e9ec175932034936b4650e840263
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3211392
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#931653}
NOKEYCHECK=True
GitOrigin-RevId: 281da3f5f055ef43123c775ab1182f26e89b7ea6