[iOS] testDismissSigninFromTabSwitcher and testCancelFromSyncOffLink
The |revokeSyncConsent| method doesn't send |OnPrimaryAccountChanged|
notifications. Therefore when |testCancelFromSyncOffLink| runs
tab switcher code didn't receive the notification that the sync consent
was revoked.
The issue is at the end of of the test, the identity is signed out,
and the tab switcher doesn't listen for sign-in consent revoke
notification.

So when |testDismissSigninFromTabSwitcher| runs, the code still thinks
the user has sync consent.

The fix is to get the identity signed in only in
|testCancelFromSyncOffLink|, using internal code instead of doing
sign-in&sync and revoke the sync consent.

Fixed: 1257107
Change-Id: If88534fbcc5f57602224023e8ca7ba807726f80a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3208047
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Jérôme <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#931693}
NOKEYCHECK=True
GitOrigin-RevId: 3991db49d9d06b519a42456aca93bbb8e70985ae