[ios][catalyst] Fix compilation of "chrome" for catalyst
The class UISheetPresentationController (and its delegate) are
not available on catalyst (even though the documentation says
it is available in macOS 15.0+, the current SDK does not define
them).

Put their use behind `#if !TARGET_OS_MACCATALYST` which is
automatically defined by the compiler when building for catalyst.

Bug: none
Change-Id: Iacf48d5692cc146c2f00d807b0d1a84d5df7c2c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3231416
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#932996}
NOKEYCHECK=True
GitOrigin-RevId: 3b94a08f64922dc24d55d001be46d110f2b1f435