[iOS][ForcedSignIn]🏢 Change access point when policy enabled
The sign-in screen mediator has its own logger internally to log
information on the sign-in access point. The access point has to be
changed when sign-in is done when the forced sign-in policy is enabled.

The access point provided when invoking
-[ApplicationCommands showSignin] isn't used for logging. The mediator
is the one doing the access point logging.

As discussed in http://shortn/_mQQ6WaIAOn and http://shortn/_buqlCJDkdu
we didn't want to introduce a new access point to take into account the
cases where the forced sign-in prompt is shown in the FRE, in which case
the access point will still be considered to be forced sign-in (hiding
the FRE context).

If we want more granularity, we will need to introduce new histograms.

Bug: 1261153
Change-Id: I3e80979a12e9ad39589c105e5d0f4d8a7480efaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3227105
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Jérôme <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#933625}
NOKEYCHECK=True
GitOrigin-RevId: ec7aaf97c8faa13b6bda4d97b6ef135ff39a9eb7