[ios] Fixed thumbstrip layout issue on iOS 15
When a reveal start, on iOS 15, the safe area insets of the BVC is
reset to empty, causing a shift when in viewSafeAreaInsetsDidChange
the tabStripView is repositioned. Rather than patching offsets in
a few places, this CL changes the use of safeAreaInsets to be
those of the root view of the BVC, which doesn't change.

Re-enabled tests that are now passing on iOS 15.

Bug: 1220137
Change-Id: I800b03e717b44fd75949ee21f6b7de50e36cd626
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3240343
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#934841}
NOKEYCHECK=True
GitOrigin-RevId: 6e5e079aa9a7c51ed0f0d5cdcca2f8fa0adaf207