breadcrumbs: rename application breadcrumbs logger
This change renames ApplicationBreadcrumbsLoggerIOS to just
ApplicationBreadcrumbsLogger, removing the IOS suffix. This is more
concise, and matches the naming style of other platform-specific
classes in the same folder, e.g., BreadcrumbManagerTabHelper.

The directory path is sufficient to inform the reader that this is the
iOS implementation of the class.

Bug: 1163762
Change-Id: I020402e32d2df3ba2633c27222e8cbf351896910
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3243189
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#935693}
NOKEYCHECK=True
GitOrigin-RevId: db3c4f57c5624d6ec1b644036527644dfca8710e