[ios] Sort WebStateObserverBridge and CRWWebStateObserver methods
Sort the WebStateObserverBridge and CRWWebStateObserver methods to
be in the same order as the methods of WebStateObserver. Sort them
in the implementation too.

Also fix naming of Objective-C method parameter to conform to the
style guide.

Bug: none
Change-Id: If3277bf94677d465d8e800d7be73126aae855801
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3250674
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#936304}
NOKEYCHECK=True
GitOrigin-RevId: edd7f09668320a501fe0d3e332e116f200a1f9b5