[ios] Ensure thumbstrips animatees are always in same order
Right now they are randomize and it could hide bugs now or later.
The code used an NSHashMap with weak references. I am moving to
and NSMutableOrderedSet which has string link, however I
checked the code and traced it in the debugger and no
memory leaks occur. The pan handler is deallocated properly
by the stop method of the ThumbStripCoordinator which is the
only user of it.

Bug: 1264349
Change-Id: Ic156d55630b5e4c9d3f87c3ee1a9c527700066d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3250510
Commit-Queue: David Jean <djean@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#936332}
NOKEYCHECK=True
GitOrigin-RevId: 404d063fd4c01a45fb0ac939767ce05c20c6bee2