[iOS] Make QRGenerator its own VC
This CL duplicates part of the logic from ConfirmationAlertVC to
QRGeneratorVC. QRGeneratorVC used to be a subclass of ConfirmationAlert.
Now it is directly inheriting from UIViewController.
This allows to simplify the logic from ConfirmationAlertVC as it
previously needed to have special logic just for QRGenerator.

Bug: None
Change-Id: Ibf419f838fa3af8eb7b704325e6facdb948ce406
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3234458
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#936438}
NOKEYCHECK=True
GitOrigin-RevId: e22f57623e9be17059fd0f1d02b9a740c8d9afe1