[iOS] Temporary fix for view hierarchy crash
We can't reproduce this crash since we don't have enough information.
This CL will check if the view hierarchy is correct, and if it is not,
it logs a DCHECK and fixes it. Therefore, it should fix the crash for
users and give us information on how to proactively fix this issue.

Bug: 1262536
Change-Id: Ib7daedbd4865f94d0bfdfea4ef2da6f7ff26236e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3254526
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#937018}
NOKEYCHECK=True
GitOrigin-RevId: 3933159ac358b67ec38d6caff7ad7b6df97f62ba