[iOS] Factor WebDelegate responsibilities out of BVC
This CL factors the WebStateDelegate responsibilities for Chrome
(previously handled by the BVC) into a browser agent. This also removes
an incomplete tab helper implementation of WebStateDelegate.

WebStateDelegateBrowserAgent is now the object assigned as the delegate
for each WebState. Since this is a browser agent, there is one of them
per browser, and thus one per WebStateList, so each web state has as its
delegate the browser agent for the web state list it is contained in.

In fact, in addition to being the WebStateDelegate, the browser agent
is also responsible for setting and unsetting the delegate on each web
sate as they are added to/removed from the web state list.

WebStateDelegateBrowserAgent requires the TabInsertionBrowserAgent to
handle tab creation. It also has and UI-dependent dependencies
for context menus and other view-related feature implementations. The
UI code (specifically the BrowserCoordinator) passes the UI
dependencies to the browser agent as soon as they are available; any
WebStateDelegate functions that require these dependencies before they
are injected are (as far as possible) silent no-ops.

To help encapsulate the context menu configuration provider, it is now
provided with a base view controller on initialization instead of
per-method-call.

Various BUILD.gn cleanups are included to avoid some dependency cycles.

Bug: 986956
Change-Id: Ia63af2a5fe3836f833be3d9525ce5d9436b58e89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3226332
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#937864}
NOKEYCHECK=True
GitOrigin-RevId: 4841677e6fe80a81055b78a653f26b244df0f27e