[iOS] Duplicate SignIn screen
This CL copies the code of the SignIn screen to the SigninSync screen.
This gives a good foundation to start building the SigninSync screen.

It also updates the screens presented by the new first run to use the
new bundled screen instead of the previous two. The tests need to be
updated to take it into account.

Fixed: 1266033
Change-Id: Ib2041d7ffee09e4aedaa7a53c10722a958eb1226
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3259167
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#938498}
NOKEYCHECK=True
GitOrigin-RevId: 7c4add16d5e31832a2f18534bffd4a585a9464cf