[iOS] Allow password form filling with disabled username element.
Password filling in a form should not be stopped when the username
element is disabled.
This CL is also removing the check for a read-only/disabled username
element value. This condition was introduced when passwords were
autofilled on pageload. This was a protection for the case when Chrome
picked wrong credential to fill. Now credentials are autofilled only on
user action, we may omit this.

Bug: 1266401
Change-Id: Ib6a8848b54f89538166e90ae8a0fca23a6ada846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3262108
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#938788}
NOKEYCHECK=True
GitOrigin-RevId: 847277bc70159894defdaa1f674d516adb29bc5f