[iOS][CPE promo] Rewrite ConsentViewController and ConsentCoordinator
ConsentViewController, which used to extend from
ConfirmationAlertViewController, changes the parent class to the new
PromoStyleViewController so that it adapts to the new promo layout.

The original ConsentViewController/Coordinator is renamed to
ConsentLegacyViewController/Coordinator. A flag check is implemented and
added to switch between the new view and the legacy view.

As a result, ConsentCoordinator now implements
PromoStyleViewControllerDelegate instead of
ConfirmationAlertActionHandler. Please check this link for the UI:
https://drive.google.com/drive/folders/1yc46DiHrIEFEFhT8XiXD3DI7v2n3AiPK?resourcekey=0-434hlGk1740gDRpnDv0lMQ&usp=sharing

The option to dismiss the view without without confirming consent is
also removed.

Bug: 1238808
Change-Id: Ia1c7bb59c1692b0f3a827e8834270c2fc2db5df7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3237388
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939211}
NOKEYCHECK=True
GitOrigin-RevId: e876fa8196f67585c3bbda0366e7d616341a8f42