[ios] Ensure returned NSIndexPath* is valid
The method -indexForCellPathWithOffset:fromPath: could return
an NSIndexPath* pointing to the first item in an empty section.
Accessing the corresponding item would result in an out-of-bound
access in an NSArray in the ListModel implementation.

Check that the section is not empty when the NSIndexPath* is
created. This should ensure that NSIndexPath* is valid.

Fixed: 1266938
Change-Id: I42e0780bf708bca93b9dfadff221fd90fa283b2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3263833
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939311}
NOKEYCHECK=True
GitOrigin-RevId: f20a246d71eccb466da55a23e44be30005b9f794