[ios] Expose API for "unrealized" WebState
Add new methods to WebState and WebStateObserver necessary to
expose the new "unrealized" state of the WebState instances.
As the feature is not yet implemented, the methods are stubs,
but this allow converting client code.

Add unit tests for the new method of WebStateObserverBridge
(and also test for the few untested methods). Sort the methods
and properties of the test objects (in the same order as the
methods of WebStateObserver).

Bug: 1264453
Change-Id: I6a8b2334940a60e6e6311f2d9ee718b75ed46778
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3260613
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939340}
NOKEYCHECK=True
GitOrigin-RevId: aa57c60b8a678a1d1fb8762dcefc76fcaab30287