[ios] Support "unrealized" WebStates in FindTabHelper
FindInPageController depends on web::FindInPageManager which is
only created when the WebState is fully realized (it is created
by the CRWWebController) thus can't be created when the WebState
is not yet realized.

Change FindTabHelper wait for the WebState to be realized before
creating the FindInPageController. This should not introduce any
behaviour change as user should be not able to interact with an
"unrealized" WebState, and thus can't use "find in page".

Stop FindInPageController from observing WebState, since it is
only listening for WebStateDestroyed() but FindTabHelper also
listen for that even and call -detachFromWebState (which is the
method that FindInPageController invokes).

Bug: 1264453
Change-Id: I78b749430e2748bb4a145ccbfcc6fd9e842fbb87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268193
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#939833}
NOKEYCHECK=True
GitOrigin-RevId: 5d671ce9e70c32510c2ee947ad3cb0569a359e81