[iOS][ForcedSignIn]🏢 Fix dismissing modals with forced sign-in prompt
Simpler version of this fix: http://crrev.com/c/3268484

This CL fixes 2 dcheck crashes that are triggered when dismissing modals
with a trigger that isn't from an interaction with the sign-in prompt:
  > The crash that happens when the
-dismissModalsWithCompletion command is called while the forced sign-in
prompt is presented. This doesn't happen when closing the app where the
signin coordinator is interrupted without dismissal.

    Problematic DCHECK: http://shortn/_krVOzQcYEu

  > The crash that happens when showing the forced sign-in prompt
    after dismissing (externally) the sign-in prompt that is presented
    on the Settings menu.

    Problematic DCHECK: http://shortn/_DmTFP8PjQI

These issues were caused by crrev.com/c/3254141 which purpose was
to fix another issue.

The problems are caused by the forced signin prompt being reshown
immediately after dismissing the sign-in prompt (any of them), which
will populate the |signinCoordinator| property of SceneController
before the process that dismisses the modal is fully done, which will
then trigger DCHECK errors because the property is populated before
the process dismissing the sign-in prompt is formally done.

A solution if to make sure that the force sign-in prompt is only
reshown after the DCHECK.

The best solution would be probably to have a signal when modals are
dismissed but this would require more design and thinking. Filled
http://crbug.com/1267900 to investigate that.

============
Visualization of the time windows of sign-in signal VS dismiss modals.

~~~
Problem - Misalignment of the time windows

The signal (represented by 'x') is fired during dismiss modals.

                YES
         |-------------|
_________|             |x--------- NO
SceneState.signinInProgress

                |-------x-----|
________________|             |----------
dismiss settings modal

---------------------------------->
time

~~~
Fix - Make sure signinInProgress covers dismiss modals

The signal (represented by 'x') is fired after dismiss modals.

                      YES
         |------------------------|
_________|                        |x--------- NO
signinInProgress

                |-------------|
________________|             |----x-----
dismiss settings modal

---------------------------------->
time
~~~~

============

============
Note about the issue with -dismissModalsWithCompletion

Normally we won't expect -dismissModalsWithCompletion to be called when
the forced sign-in prompt is presented, but this can happen when testing
where the command can be executed while forced sign-in prompt blocks
the UI (I think normally this wouldn't be allowed).

See crrev.com/c/3267844 as example of a context where
-dismissModalsWithCompletion may be called while the forced sign-in
prompt is presented.
============

Bug: 1267723
Change-Id: Ifbe427840aca3082b19fd570e8862ea04a79b046
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3264702
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940281}
NOKEYCHECK=True
GitOrigin-RevId: a7fd16319ac4208c392e844015853e79129dd0b9